fix imports in app
