+++ frontend/index.html
@@ -26,6 +26,9 @@ function App() {
   const holisticRef = useRef(null);
   const [transcript, setTranscript] = useState('');
   const [showLms, setShowLms] = useState(false);
+  const [status, setStatus] = useState('Connecting...');
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState('');
 
   useEffect(() => {
     async function init() {
@@ -33,15 +36,31 @@ function App() {
       videoRef.current.srcObject = stream;
 
       socketRef.current = new WebSocket('ws://' + location.hostname + ':8000/ws');
-      recorderRef.current = new MediaRecorder(stream, {mimeType: 'video/webm'});
-      recorderRef.current.ondataavailable = ev => {
-        if (ev.data.size > 0 && socketRef.current.readyState === 1) {
-          socketRef.current.send(ev.data);
-        }
+      socketRef.current.onopen = () => {
+        setStatus('Connected');
+        setError('');
+      };
+      socketRef.current.onclose = () => {
+        setStatus('Disconnected');
+        setError('WebSocket disconnected');
+        setLoading(false);
+      };
+      socketRef.current.onerror = () => {
+        setError('WebSocket error');
       };
       socketRef.current.onmessage = ev => {
         const data = JSON.parse(ev.data);
         setTranscript(data.transcript || '');
+        setLoading(false);
+        if (data.error) setError(data.error); else setError('');
+      };
+
+      recorderRef.current = new MediaRecorder(stream, {mimeType: 'video/webm'});
+      recorderRef.current.ondataavailable = ev => {
+        if (ev.data.size > 0 && socketRef.current.readyState === 1) {
+          setLoading(true);
+          socketRef.current.send(ev.data);
+        }
       };
       recorderRef.current.start(2000);
 
@@ -87,6 +106,9 @@ function App() {
       React.createElement('video', {ref: videoRef, autoPlay: true, playsInline: true}),
       React.createElement('canvas', {ref: canvasRef, width: 320, height: 240})
     ),
+    React.createElement('p', null, 'Status: ' + status),
+    loading ? React.createElement('p', null, 'Transcribing...') : null,
+    error ? React.createElement('p', {style: {color: 'red'}}, error) : null,
     React.createElement('p', null, transcript),
     React.createElement('label', null,
       React.createElement('input', {type: 'checkbox', checked: showLms,
